// src/models/companies/types/CompanyHasBrandRelationship.ts
import type {BaseNode} from "../../../controllers/nodes/types/BaseNode"

export type CompanyHasBrandRelationship = {
    company_id: number
    brand_id: number
    relationship_id: number
    relationship_name: string
    relationship_partner?: BaseNode
    created_at: string
    updated_at: string
}